# api/openapi.yaml
openapi: 3.1.0
info:
  title: API
  description: >
    API for managing statistics, organizations, teams, and user access control.
  version: 0.1.0
servers:
  - url: http://localhost:3000/api/v1
    description: Development server
components:
  securitySchemes:
    AccountAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token for account-level operations.
    OrganizationAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Authentication token for organization-specific operations.
  schemas:
    Registration:
      type: object
      description: Data required to register a new user account.
      properties:
        email: { type: string, format: email, description: "The user's email address." }
        password: { type: string, format: password, description: "The user's password." }
        firstName: { type: string, description: "The user's first name." }
        lastName: { type: string, description: "The user's last name." }
    Login:
      type: object
      description: Credentials required to authenticate an existing user.
      properties:
        email: { type: string, format: email, description: "The registered email address." }
        password: { type: string, format: password, description: "The account password." }
...
    OrganizationDetails:
      type: object
      description: Full details of an organization.
      properties:
        id: { type: string, format: uuid, description: "Unique identifier." }
        ownerId:
          { type: string, format: uuid, description: "Identifier of the organization owner." }
        name: { type: string, description: "Official name." }
        slug: { type: string, description: "URL slug." }
        avatarUrl: { type: string, format: uri, description: "URL to the organization's avatar." }
        createdAt: { type: string, format: date-time, description: "Creation timestamp." }
        updatedAt: { type: string, format: date-time, description: "Last update timestamp." }
    Member:
      type: object
      description: A member within an organization.
      properties:
        id: { type: string, format: uuid, description: "Unique membership ID." }
        accountId:
          { type: string, format: uuid, description: "The global account ID of the member." }
        email: { type: string, format: email, description: "Member's email address." }
        firstName: { type: string, description: "Member's first name." }
        lastName: { type: string, description: "Member's last name." }
        roles:
          type: array
          description: "List of role IDs assigned to the member."
          items: { type: string, format: uuid }
        groups:
          type: array
          description: "List of group IDs the member belongs to."
          items: { type: string, format: uuid }
        status:
          {
            type: string,
            enum: [active, invited, suspended],
            description: "Current membership status.",
          }
        joinedAt: { type: string, format: date-time, description: "Timestamp of joining." }
...
    Role:
      type: object
      description: A role defining a set of permissions.
      properties:
        id: { type: string, format: uuid, description: "Unique role ID." }
        name: { type: string, description: "Role name." }
        isSystem:
          { type: boolean, description: "Indicates if the role is a protected system role." }
        permissions:
          type: array
          description: "Permission keys associated with this role."
          items: { type: string }
        createdAt: { type: string, format: date-time, description: "Creation timestamp." }
        updatedAt: { type: string, format: date-time, description: "Last update timestamp." }
...
paths:
  /auth/register:
    post:
      tags: ["Auth"]
      summary: Register a new account
      description: Registers a user with email, password, and personal details.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Registration" }
      responses:
        "204":
          description: Successfully created a new account
  /auth/login:
    post:
      tags: ["Auth"]
      summary: Login to an account
      description: Authenticates a user and returns necessary tokens or session state.
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Login" }
      responses:
        "204":
          description: Successfully authenticated to an account
...
  /auth/refresh:
    post:
      tags: ["Auth"]
      summary: Issue new access token
      description: Refreshes the session and issues a new access token for the organization context.
      security: [{ OrganizationAuth: [] }]
      responses:
        "204":
          description: Successfully issued a new access token
  /auth/logout:
    post:
      tags: ["Auth"]
      summary: Logout from an account
      description: Invalidates current session or tokens.
      security: [{ AccountAuth: [] }]
      responses:
        "204":
          description: Successfully logged out of an account
...
  /organizations/{organization_id}:
    get:
      tags: ["Organizations"]
      summary: Get organization details
      description: Fetches comprehensive data for a single specific organization.
      security: [{ OrganizationAuth: [] }]
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Returns details of a specific organization
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrganizationDetails" }
  /organizations/{organization_id}/members:
    get:
      tags: ["Members"]
      summary: Get organization members
      description: Lists all members currently in the organization.
      security: [{ OrganizationAuth: [] }]
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Returns a list of organization members
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Member" }
    post:
      tags: ["Members"]
      summary: Adds an organization member
      description: Invites or adds a new member to the organization.
      security: [{ OrganizationAuth: [] }]
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully added new member to organization
...
  /organizations/{organization_id}/roles:
    get:
      tags: ["Roles"]
      summary: Get organization roles
      description: Lists all defined roles in the organization.
      security: [{ OrganizationAuth: [] }]
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Returns a list of organization roles
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Role" }
    post:
      tags: ["Roles"]
      summary: Create an organization role
      description: Defines a new role with specific permissions.
      security: [{ OrganizationAuth: [] }]
      parameters:
        - name: organization_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Successfully created a role
...